ci: cargo fmt + split Main/Kernels table out of main_tab for the 500-LoC cap
app.rs back to 500 with a tighter toast-TTL comment; draw_kernels moves to
main_tab_kernels.rs unchanged (its tests already drive it through
render::draw). The benchmark-gate failure is the standing ~8 GPU-hour gate,
same as #405/#395.
